huge refactoring/optimization: * optimize memory usage (int replace with byte) * reworked find cell's depth - was very slow * remove redundant conversions byte-int-byte * add highload v3 bulk transfer tests todo: * replace old hash() in Cell * rework call of calculateHashes() for static boc deserialization() * cleanup

+++ bitstring/src/test/java/org/ton/java/bitstring/TestBitString.java
@@ -187,26 +187,26 @@ public class TestBitString {
     public void testBitStringGetTopUppedArray() {
         BitString bitString0 = new BitString(8);
         bitString0.writeUint(BigInteger.valueOf(200), 8);
-        int[] b0 = bitString0.getTopUppedArray();
+        byte[] b0 = bitString0.getTopUppedArray();
         int[] result0 = {200}; // (-56 & 0xff) = 200
         assertThat(b0).isEqualTo(result0);
 
         BitString bitString1 = new BitString(32);
         bitString1.writeUint(BigInteger.valueOf(200), 32);
-        int[] b1 = bitString1.getTopUppedArray();
+        byte[] b1 = bitString1.getTopUppedArray();
         int[] result1 = {0, 0, 0, 200};
         assertThat(b1).isEqualTo(result1);
 
         BitString bitString2 = new BitString(32);
         bitString2.writeCoins(BigInteger.TEN);
-        int[] b2 = bitString2.getTopUppedArray();
+        byte[] b2 = bitString2.getTopUppedArray();
         int[] result2 = {16, 168};
         assertThat(b2).isEqualTo(result2);
 
         BitString bitString3 = new BitString(1023);
         bitString3.writeCoins(BigInteger.TEN);
         bitString3.writeUint(333L, 64);
-        int[] b3 = bitString3.getTopUppedArray();
+        byte[] b3 = bitString3.getTopUppedArray();
         int[] result3 = {16, 160, 0, 0, 0, 0, 0, 0, 20, 216};
         assertThat(b3).isEqualTo(result3);
     }
